Added stamina and mana potions

+++ Assets/Scripts/World/Inventory/ItemTypesData/PotionsData/ManaPotionItemData.cs
@@ -5,6 +5,6 @@ namespace World.Inventory.ItemTypesData.PotionsData
     [CreateAssetMenu(fileName = "ManaPotionItemData", menuName = "Data/Inventory Data/Potion/Mana")]
     public class ManaPotionItemData : PotionItemData
     {
-        public int manaPercent;
+        public float manaPercent;
     }
 }
